P#58329 Bundle settings on one page Remove admin page modules
