+++ plugin/Controller/AdminViewController.php
@@ -181,12 +181,6 @@ class AdminViewController
 		add_action( 'load-'.$hookEditForm, array($this->_pAdminPageFormSettings, 'handleAdminNotices'));
 		add_action( 'load-'.$hookEditForm, array($this->_pAdminPageFormSettings, 'checkForms'));
 
-		// Modules
-		$pAdminPageModules = new AdminPageModules($this->_pageSlug);
-		add_submenu_page( $this->_pageSlug, __('Modules', 'onoffice-for-wp-websites'), __('Modules', 'onoffice-for-wp-websites'),
-			$roleModules, $this->_pageSlug.'-modules', array($pAdminPageModules, 'render'));
-		add_action( 'admin_init', array($pAdminPageModules, 'registerForms'));
-
 		// Estates: edit list view (hidden page)
 		$hookEditList = add_submenu_page(null, null, null, $roleEstate, $this->_pageSlug.'-editlistview',
 			array($this->_pAdminListViewSettings, 'render'));
