+++ plugin/Gui/AdminPageApiSettings.php
@@ -24,6 +24,7 @@ namespace onOffice\WPlugin\Gui;
 use onOffice\WPlugin\Model\FormModel;
 use onOffice\WPlugin\Model\InputModelOption;
 use onOffice\WPlugin\Renderer\InputModelRenderer;
+use onOffice\WPlugin\Types\MapProvider;
 use function __;
 use function admin_url;
 use function do_settings_sections;
@@ -57,7 +58,9 @@ class AdminPageApiSettings
 		parent::__construct($pageSlug);
 		$this->_encrypter = $this->getContainer()->make(SymmetricEncryption::class);
 		$this->addFormModelAPI();
+        $this->addFormModelMapProvider($pageSlug);
 		$this->addFormModelGoogleCaptcha();
+        $this->addFormModelDetailView($pageSlug);
 		$this->addFormModelGoogleMapsKey();
 		$this->addFormModelGoogleBotSettings();
 	}
@@ -275,4 +278,59 @@ class AdminPageApiSettings
 
 		printf('<div class="%1$s"><p>%2$s</p></div>', esc_attr($class), esc_html($message));
 	}
+
+    /**
+     *
+     * @param string $pageSlug
+     *
+     */
+
+    private function addFormModelMapProvider(string $pageSlug)
+    {
+        $groupSlugMaps = 'onoffice-maps';
+        $mapProviderLabel = __('Map Provider', 'onoffice-for-wp-websites');
+        $pInputModelMapProvider = new InputModelOption($groupSlugMaps, 'mapprovider',
+            $mapProviderLabel, InputModelOption::SETTING_TYPE_NUMBER);
+        $pInputModelMapProvider->setHtmlType(InputModelOption::HTML_TYPE_RADIO);
+        $selectedValue = get_option($pInputModelMapProvider->getIdentifier(), MapProvider::PROVIDER_DEFAULT);
+        $pInputModelMapProvider->setValue($selectedValue);
+        $pInputModelMapProvider->setValuesAvailable([
+            MapProvider::OPEN_STREET_MAPS => __('OpenStreetMap', 'onoffice-for-wp-websites'),
+            MapProvider::GOOGLE_MAPS => __('Google Maps', 'onoffice-for-wp-websites'),
+        ]);
+
+        $pFormModel = new FormModel();
+        $pFormModel->addInputModel($pInputModelMapProvider);
+        $pFormModel->setGroupSlug($groupSlugMaps);
+        $pFormModel->setPageSlug($pageSlug);
+        $pFormModel->setLabel(__('Maps', 'onoffice-for-wp-websites'));
+
+        $this->addFormModel($pFormModel);
+    }
+
+    /**
+     *
+     * @param string $pageSlug
+     *
+     */
+    private function addFormModelDetailView(string $pageSlug)
+    {
+        $groupSlugView = 'onoffice-detail-view';
+        $showTitleInUrl = __('Show title in URL', 'onoffice-for-wp-websites');
+        $pInputModelShowTitleUrl = new InputModelOption($groupSlugView, 'showTitleUrl',
+            $showTitleInUrl, InputModelOption::SETTING_TYPE_BOOLEAN);
+        $pInputModelShowTitleUrl->setHtmlType(InputModelOption::HTML_TYPE_CHECKBOX);
+        $pInputModelShowTitleUrl->setValuesAvailable(1);
+        $pInputModelShowTitleUrl->setValue(get_option($pInputModelShowTitleUrl->getIdentifier()) == 1);
+        $pInputModelShowTitleUrl->setDescriptionTextHTML('If this checkbox is selected, the title of the property will be part of the URLs of the detail views.
+		The title is placed after the record number, e.g. <code>/1234-nice-location-with-view</code>. No more than the first five words of the title are used.');
+
+        $pFormModel = new FormModel();
+        $pFormModel->addInputModel($pInputModelShowTitleUrl);
+        $pFormModel->setGroupSlug($groupSlugView);
+        $pFormModel->setPageSlug($pageSlug);
+        $pFormModel->setLabel(__('Detail View URLs', 'onoffice-for-wp-websites'));
+
+        $this->addFormModel($pFormModel);
+    }
 }
